fix comments and add a test

+++ contracts/MelonFarm.sol
@@ -10,22 +10,21 @@ pragma solidity 0.4.24;
 import './Meloncoin.sol';
 
 /**
- * @title: MelonFarm deploys and manages meloncoins
+ * @title MelonFarm deploys and manages meloncoins
  */
 contract MelonFarm {
-
   Meloncoin[] public deployed;
 
 /**
- * @dev: Creates a new Meloncoin, assigning all tokens to the sender
- * @param uint _initialSupply : How many melon seeds are planted
- * @param uint _plantDate : When the melons are planted
- * @param uint8 _growingPeriod : How long the melons take to grow in days
- * @param uint8 _ripePeriod : The shelf life of the melons in days. This is the
+ * @dev Creates a new Meloncoin, assigning all tokens to the sender
+ * @param _initialSupply : How many melon seeds are planted
+ * @param _plantDate : When the melons are planted
+ * @param _growingPeriod : How long the melons take to grow in days
+ * @param _ripePeriod : The shelf life of the melons in days. This is the
  * period in which meloncoin can be redeemed for an investment-grade melon
  * @return : a new Meloncoin contract
  */
-  function launchMeloncoin(uint _initialSupply,
+  function launchMeloncoin(uint8 _initialSupply,
 			   uint _plantDate,
 			   uint8 _growingPeriod,
 			   uint8 _ripePeriod) public returns (Meloncoin) {
